refactored index.ts
